Added the solution for the first task
