--- homework-09/cmd/main.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-	"go-course-4/homework-09/pkg/maxage1"
-	"go-course-4/homework-09/pkg/maxage2"
-)
-
-func main() {
-	maxage1.MaxAge()
-	maxage2.MaxAge()
-}
